Replace use of UdpClient with Socket.
Using Connect() as a convenience to cache the DNS lookup has the
unfortunate side effect of making the socket interface bound. The send
would therefore not recover when the network interface changes.

Do a manual DNS lookup initially (as for all other platforms) and use
a bound low level socket instead.